requirements.txt
google-cloud-secret-manager
langgraph 
langchain 
python-dotenv
google-genai
langchain-google-genai
google-auth